backend/requirements.txt
# FastAPI & Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
starlette>=0.40.0

# Database
SQLModel>=0.0.22
psycopg[binary]>=3.2.1
alembic>=1.13.2

# Cache & Storage
redis>=5.0.7
boto3>=1.35.0

# Auth & Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Config & Utils
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.27.0

# AI
anthropic>=0.18.0

# Excel
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
